refactor: split Actions components
